Create AbstractSuggestBox class
